use signed
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@48163 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/shared/wlutils.h
@@ -342,7 +342,7 @@ struct site_survey_list {
 	uint8 dtim_period;	/* DTIM period */
 	unsigned long long active;	/* channel active time */
 	unsigned long long busy;	/* channel busy time */
-	uint8 numsta;
+	int16 numsta;
 };
 
 struct wifi_interface {
